Move print function to a utils.py source file
